Crate simple_chess

This is rust chess library focused on simplicity. Speed is not main focus. This library isn’t good for chess engine or AI. This library is good for making simple chess game.

Examples

Read README.md for more info.

Structs

Board

Board stores position and history of position. Position is represent by array of Option<Piece>, with unchangeable size of 8x8.

ChessMove

Represent a chess move, stores info about move (start, destination, promotion).

Piece

Represent a piece with type and color. Doesn’t hold any information about their position on the crate::Board.

Square

Represent a index on the Board.

Enums

Color

File

File enum for simple orientation on the board

GenMask

Masks represents differnet types of generating moves via gen_move() function.

PieceType

represent type of piece

Rank

rank enum for simple orientation on the board

Functions

gen_moves

Create valid moves from Board, depends on GenMask.