Struct simple_chess::Square

pub struct Square(_);

Represent a index on the Board.

Implementations

impl Square

pub fn up(
&self,
mul: i32,
promo: Option<PieceType>
) -> Result<ChessMove, &'static str>

Create ChessMove from Square to n squares up.

pub fn down(
&self,
mul: i32,
promo: Option<PieceType>
) -> Result<ChessMove, &'static str>

Create ChessMove from Square to n squares down.

pub fn right(
&self,
mul: i32,
promo: Option<PieceType>
) -> Result<ChessMove, &'static str>

Create ChessMove from Square to n squares right.

pub fn left(
&self,
mul: i32,
promo: Option<PieceType>
) -> Result<ChessMove, &'static str>

Create ChessMove from Square to n squares left.

pub fn down_right(
&self,
mul: i32,
promo: Option<PieceType>
) -> Result<ChessMove, &'static str>

Create ChessMove from Square to n squares right and down.

pub fn down_left(
&self,
mul: i32,
promo: Option<PieceType>
) -> Result<ChessMove, &'static str>

Create ChessMove from Square to n squares left and down.

pub fn up_right(
&self,
mul: i32,
promo: Option<PieceType>
) -> Result<ChessMove, &'static str>

Create ChessMove from Square to n squares right and up.

pub fn up_left(
&self,
mul: i32,
promo: Option<PieceType>
) -> Result<ChessMove, &'static str>

Create ChessMove from Square to n squares left and up.

pub const A1: Square = _

built in constant A1 Square for easy indexing to Board

pub const B1: Square = _

built in constant B1 Square for easy indexing to Board

pub const C1: Square = _

built in constant C1 Square for easy indexing to Board

pub const D1: Square = _

built in constant D1 Square for easy indexing to Board

pub const E1: Square = _

built in constant E1 Square for easy indexing to Board

pub const F1: Square = _

built in constant F1 Square for easy indexing to Board

pub const G1: Square = _

built in constant G1 Square for easy indexing to Board

pub const H1: Square = _

built in constant H1 Square for easy indexing to Board

pub const A2: Square = _

built in constant A2 Square for easy indexing to Board

pub const B2: Square = _

built in constant B2 Square for easy indexing to Board

pub const C2: Square = _

built in constant C2 Square for easy indexing to Board

pub const D2: Square = _

built in constant D2 Square for easy indexing to Board

pub const E2: Square = _

built in constant E2 Square for easy indexing to Board

pub const F2: Square = _

built in constant F2 Square for easy indexing to Board

pub const G2: Square = _

built in constant G2 Square for easy indexing to Board

pub const H2: Square = _

built in constant H2 Square for easy indexing to Board

pub const A3: Square = _

built in constant A3 Square for easy indexing to Board

pub const B3: Square = _

built in constant B3 Square for easy indexing to Board

pub const C3: Square = _

built in constant C3 Square for easy indexing to Board

pub const D3: Square = _

built in constant D3 Square for easy indexing to Board

pub const E3: Square = _

built in constant E3 Square for easy indexing to Board

pub const F3: Square = _

built in constant F3 Square for easy indexing to Board

pub const G3: Square = _

built in constant G3 Square for easy indexing to Board

pub const H3: Square = _

built in constant H3 Square for easy indexing to Board

pub const A4: Square = _

built in constant A4 Square for easy indexing to Board

pub const B4: Square = _

built in constant B4 Square for easy indexing to Board

pub const C4: Square = _

built in constant C4 Square for easy indexing to Board

pub const D4: Square = _

built in constant D4 Square for easy indexing to Board

pub const E4: Square = _

built in constant E4 Square for easy indexing to Board

pub const F4: Square = _

built in constant F4 Square for easy indexing to Board

pub const G4: Square = _

built in constant G4 Square for easy indexing to Board

pub const H4: Square = _

built in constant H4 Square for easy indexing to Board

pub const A5: Square = _

built in constant A5 Square for easy indexing to Board

pub const B5: Square = _

built in constant B5 Square for easy indexing to Board

pub const C5: Square = _

built in constant C5 Square for easy indexing to Board

pub const D5: Square = _

built in constant D5 Square for easy indexing to Board

pub const E5: Square = _

built in constant E5 Square for easy indexing to Board

pub const F5: Square = _

built in constant F5 Square for easy indexing to Board

pub const G5: Square = _

built in constant G5 Square for easy indexing to Board

pub const H5: Square = _

built in constant H5 Square for easy indexing to Board

pub const A6: Square = _

built in constant A6 Square for easy indexing to Board

pub const B6: Square = _

built in constant B6 Square for easy indexing to Board

pub const C6: Square = _

built in constant C6 Square for easy indexing to Board

pub const D6: Square = _

built in constant D6 Square for easy indexing to Board

pub const E6: Square = _

built in constant E6 Square for easy indexing to Board

pub const F6: Square = _

built in constant F6 Square for easy indexing to Board

pub const G6: Square = _

built in constant G6 Square for easy indexing to Board

pub const H6: Square = _

built in constant H6 Square for easy indexing to Board

pub const A7: Square = _

built in constant A7 Square for easy indexing to Board

pub const B7: Square = _

built in constant B7 Square for easy indexing to Board

pub const C7: Square = _

built in constant C7 Square for easy indexing to Board

pub const D7: Square = _

built in constant D7 Square for easy indexing to Board

pub const E7: Square = _

built in constant E7 Square for easy indexing to Board

pub const F7: Square = _

built in constant F7 Square for easy indexing to Board

pub const G7: Square = _

built in constant G7 Square for easy indexing to Board

pub const H7: Square = _

built in constant H7 Square for easy indexing to Board

pub const A8: Square = _

built in constant A8 Square for easy indexing to Board

pub const B8: Square = _

built in constant B8 Square for easy indexing to Board

pub const C8: Square = _

built in constant C8 Square for easy indexing to Board

pub const D8: Square = _

built in constant D8 Square for easy indexing to Board

pub const E8: Square = _

built in constant E8 Square for easy indexing to Board

pub const F8: Square = _

built in constant F8 Square for easy indexing to Board

pub const G8: Square = _

built in constant G8 Square for easy indexing to Board

pub const H8: Square = _

built in constant H8 Square for easy indexing to Board