Crate simple_chess

This is rust chess library focused on simplicity. Speed is not main focus. This library is not good for chess engine or AI. This library is good for making simple chess game.

Examples

Board::deafult();
Timer::new(
    white: Duration::from_secs(600),
    black: Duration::from_secs(600),
    white_move: Duration::from_secs(0),
    white_move: Duration::from_secs(0),
    white_exten: Duration::from_secs(0),
    white_exten: Duration::from_secs(0),
);
 
Game::new(&mut Board, &mut Timer);

Read README.md for more info.

Structs

Board

Board stores position and history of position position is represent by 2D array of Squares with unchangeable size of 8x8 Squares

Game

Game struct is for managing board and timer at once

Piece

represent a piece with type and color doesn’t hold any information about their position on board

Square

represent a part index on board

Enums

Color

represent a player or color of their pieces

File

File enum for simple orientation on the board

PieceType

represent type of piece

Rank

rank enum for simple orientation on the board

Constants

A1

A2

A3

A4

A5

A6

A7

A8

B1

B2

B3

B4

B5

B6

B7

B8

C1

C2

C3

C4

C5

C6

C7

C8

D1

D2

D3

D4

D5

D6

D7

D8

E1

E2

E3

E4

E5

E6

E7

E8

F1

F2

F3

F4

F5

F6

F7

F8

G1

G2

G3

G4

G5

G6

G7

G8

H1

H2

H3

H4

H5

H6

H7

H8